Fix DatePicker unit tests
