+++ lib/src/__tests__/e2e/DatePicker.test.tsx
@@ -1,19 +1,77 @@
 import { ReactWrapper } from 'enzyme';
 import * as React from 'react';
-import DatePicker, { BaseDatePickerProps } from '../../DatePicker/DatePickerRoot';
+import DatePicker, { DatePickerProps } from '../../DatePicker/DatePicker';
 import { mount, utilsToUse } from '../test-utils';
 
-describe('e2e - DatePicker', () => {
-  let component: ReactWrapper<BaseDatePickerProps>;
+describe('e2e - DatePicker default year format', () => {
+  let component: ReactWrapper<DatePickerProps>;
   const onChangeMock = jest.fn();
+  const date = utilsToUse.date('2018-01-01T00:00:00.000Z');
+
+  beforeEach(() => {
+    component = mount(<DatePicker onChange={onChangeMock} value={date} views={['year']} />);
+  });
+
+  it('Should use year format by default for year only view', () => {
+    expect(component.find('input').props().value).toBe(
+      utilsToUse.format(date, utilsToUse.yearFormat)
+    );
+  });
+});
+
+describe('e2e - DatePicker default year month format', () => {
+  let component: ReactWrapper<DatePickerProps>;
+  const onChangeMock = jest.fn();
+  const date = utilsToUse.date('2018-01-01T00:00:00.000Z');
+
+  beforeEach(() => {
+    component = mount(
+      <DatePicker onChange={onChangeMock} value={date} views={['year', 'month']} />
+    );
+  });
+
+  it('Should use year month format by default for year & month views', () => {
+    expect(component.find('input').props().value).toBe(
+      utilsToUse.format(date, utilsToUse.yearMonthFormat)
+    );
+  });
+});
+
+describe('e2e - DatePicker default year month day format', () => {
+  let component: ReactWrapper<DatePickerProps>;
+  const onChangeMock = jest.fn();
+  const date = utilsToUse.date('2018-01-01T00:00:00.000Z');
+
+  beforeEach(() => {
+    component = mount(
+      <DatePicker onChange={onChangeMock} value={date} views={['year', 'month', 'day']} />
+    );
+  });
+
+  it('Should use default for year & month & day views', () => {
+    expect(component.find('input').props().value).toBe(
+      utilsToUse.format(date, utilsToUse.dateFormat)
+    );
+  });
+});
+
+describe('e2e - DatePicker inline variant', () => {
+  let component: ReactWrapper<DatePickerProps>;
+  const onChangeMock = jest.fn();
+  const onCloseMock = jest.fn();
+  const onOpenMock = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
     component = mount(
       <DatePicker
+        autoOk
+        variant="inline"
         animateYearScrolling={false}
-        date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
         onChange={onChangeMock}
+        onClose={onCloseMock}
+        onOpen={onOpenMock}
+        value={utilsToUse.date('2018-01-01T00:00:00.000Z')}
       />
     );
   });
@@ -22,126 +80,30 @@ describe('e2e - DatePicker', () => {
     expect(component).toBeTruthy();
   });
 
-  it('Should render proper count of days', () => {
-    expect(component.find('Day').length).toBe(35);
+  it('Should open modal with picker on click', () => {
+    component.find('input').simulate('click');
+
+    expect(component.find('WithStyles(Popover)').props().open).toBeTruthy();
+    expect(onOpenMock).toHaveBeenCalled();
   });
 
-  it('Should dispatch onChange on day click', () => {
+  it('Should close on popover close request', () => {
+    const popoverOnClose = component.find('WithStyles(Popover)').prop('onClose') as () => void;
+
+    popoverOnClose();
+
+    expect(component.find('WithStyles(Popover)').props().open).toBeFalsy();
+    expect(onCloseMock).toHaveBeenCalled();
+  });
+
+  it('Should dispatch onChange and close on day select', () => {
+    component.find('input').simulate('click');
     component
       .find('Day button')
-      .at(2)
-      .simulate('click');
-    expect(onChangeMock).toHaveBeenCalled();
-  });
-
-  it('Should render year selection', () => {
-    component
-      .find('ToolbarButton')
-      .first()
-      .simulate('click');
-
-    expect(component.find('Year').length).toBe(201);
-
-    component
-      .find('Year')
-      .at(1)
-      .simulate('click');
-    expect(onChangeMock).toHaveBeenCalled();
-  });
-});
-
-describe('e2e -- DatePicker views year', () => {
-  const onChangeMock = jest.fn();
-  const onYearChangeMock = jest.fn();
-
-  let component: ReactWrapper<BaseDatePickerProps>;
-
-  beforeEach(() => {
-    component = mount(
-      <DatePicker
-        date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
-        onChange={onChangeMock}
-        onYearChange={onYearChangeMock}
-        views={['year']}
-      />
-    );
-  });
-
-  it('Should render year selection and select year', () => {
-    expect(component.find('Year').length).toBe(201);
-
-    component
-      .find('Year')
-      .at(1)
+      .at(10)
       .simulate('click');
 
     expect(onChangeMock).toHaveBeenCalled();
-    expect(onYearChangeMock).toHaveBeenCalled();
-  });
-});
-
-describe('e2e -- DatePicker views year and month', () => {
-  const onChangeMock = jest.fn();
-  const onMonthChangeMock = jest.fn();
-
-  let component: ReactWrapper<BaseDatePickerProps>;
-
-  beforeEach(() => {
-    component = mount(
-      <DatePicker
-        date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
-        onChange={onChangeMock}
-        onMonthChange={onMonthChangeMock}
-        views={['year', 'month']}
-      />
-    );
-  });
-
-  it('Should render month selection', () => {
-    expect(component.find('Month').length).toBe(12);
-  });
-
-  it('Should switch to year selection and back to month', () => {
-    component
-      .find('ToolbarButton')
-      .first()
-      .simulate('click');
-
-    const year = component.find('Year');
-    expect(component.find('Year').length).toBe(201);
-
-    year.first().simulate('click');
-
-    expect(component.find('Month').length).toBe(12);
-  });
-
-  it('Should select month', () => {
-    component
-      .find('Month')
-      .first()
-      .simulate('click');
-
-    expect(onChangeMock).toHaveBeenCalled();
-    expect(onMonthChangeMock).toHaveBeenCalled();
-  });
-});
-
-describe('e2e -- DatePicker views year and month open from year', () => {
-  const onChangeMock = jest.fn();
-  let component: ReactWrapper<BaseDatePickerProps>;
-
-  beforeEach(() => {
-    component = mount(
-      <DatePicker
-        date={utilsToUse.date('2018-01-01T00:00:00.000Z')}
-        onChange={onChangeMock}
-        views={['year', 'month']}
-        openToYearSelection
-      />
-    );
-  });
-
-  it('Should render year selection', () => {
-    expect(component.find('Year').length).toBe(201);
+    expect(component.find('WithStyles(Popover)').props().open).toBeFalsy();
   });
 });
